Refactored all code into 1 file. Refatored the point spend algorithm too
